Move test to sendValidationEmailControllerTest
